requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==3.2
django-crispy-forms==1.14.0
django-croppie==1.3
django-debug-toolbar==3.2.4
django-htmx==1.9.0
Pillow==9.1.0
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
